refactor: move Global IFW rule models and preview providers to feature module
- Relocate `GlobalIfwRuleModels.kt` from `core:model` to `feature:globalifwrule:impl` as `GlobalIfwRuleUiModels.kt`
- Move `GlobalIfwRulePreviewParameterProvider.kt` from `core:ui` to `feature:globalifwrule:impl`
- Update all internal references and imports to point to the new package structure in the `feature` module
- Change `core:model` dependency in `core:ifw-api` from `api` to `implementation`
- Update `IfwRuleTreeEditor` preview to use explicit node definitions instead of the moved preview data
- Update copyright years to 2026 in touched files
