+++ feature/globalifwrule/impl/src/main/kotlin/com/merxury/blocker/feature/globalifwrule/impl/GlobalIfwRuleViewModel.kt
@@ -22,22 +22,22 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.merxury.blocker.core.data.respository.component.ComponentRepository
-import com.merxury.blocker.core.model.data.AdvancedGlobalIfwRuleDraft
-import com.merxury.blocker.core.model.data.GlobalIfwRuleEditMode
-import com.merxury.blocker.core.model.data.GlobalIfwRuleEditorUiState
-import com.merxury.blocker.core.model.data.GlobalIfwRuleScreenState
-import com.merxury.blocker.core.model.data.GlobalIfwRuleUiState
-import com.merxury.blocker.core.model.data.PackageRuleGroup
-import com.merxury.blocker.core.model.data.RuleItemUiState
-import com.merxury.blocker.core.model.data.SimpleGlobalIfwRuleDraft
-import com.merxury.blocker.core.model.data.SimpleRuleComponentUiState
-import com.merxury.blocker.core.model.data.SimpleTargetMode
 import com.merxury.blocker.feature.globalifwrule.impl.components.flattenComponentName
 import com.merxury.blocker.feature.globalifwrule.impl.components.toComponentType
 import com.merxury.blocker.feature.globalifwrule.impl.components.toDetailUiState
 import com.merxury.blocker.feature.globalifwrule.impl.components.toIfwRule
 import com.merxury.blocker.feature.globalifwrule.impl.components.toIfwRuleOrNull
 import com.merxury.blocker.feature.globalifwrule.impl.components.toRuleItemUiState
+import com.merxury.blocker.feature.globalifwrule.impl.model.AdvancedGlobalIfwRuleDraft
+import com.merxury.blocker.feature.globalifwrule.impl.model.GlobalIfwRuleEditMode
+import com.merxury.blocker.feature.globalifwrule.impl.model.GlobalIfwRuleEditorUiState
+import com.merxury.blocker.feature.globalifwrule.impl.model.GlobalIfwRuleScreenState
+import com.merxury.blocker.feature.globalifwrule.impl.model.GlobalIfwRuleUiState
+import com.merxury.blocker.feature.globalifwrule.impl.model.PackageRuleGroup
+import com.merxury.blocker.feature.globalifwrule.impl.model.RuleItemUiState
+import com.merxury.blocker.feature.globalifwrule.impl.model.SimpleGlobalIfwRuleDraft
+import com.merxury.blocker.feature.globalifwrule.impl.model.SimpleRuleComponentUiState
+import com.merxury.blocker.feature.globalifwrule.impl.model.SimpleTargetMode
 import com.merxury.core.ifw.IIntentFirewall
 import com.merxury.core.ifw.editor.IfwEditorNode
 import com.merxury.core.ifw.model.IfwComponentType
